Add XDoc-Store
Various playgrounds exist within Gtoolkit, e.g. the Gt Playground,
playground tabs for PharoLink and JSLink, and the evaluation windows in
the Gt inspector.

Create a general purpose xdoc store that stores timestamped versions of
the documents and supports flexible keys. The default key will be the
class name of the object using the store.

Currently accessors are provided only for retrieving the latest version,
either by modified timestamp or the timestamp encoded in the filename.

Fixes: https://github.com/feenkcom/gtoolkit/issues/952